requirements.txt
cloudshell-automation-api>=2022.1,<2023.2
attrs>=22.0,<23.2
cloudshell-shell-core~=6.0
retrying~=1.3.3
requests>=2.28,<2.32
websocket-client~=1.8.0
cloudshell-cp-core~=2.5
cloudshell-logging~=2.1
cloudshell-shell-connectivity-flow~=4.2
cloudshell-shell-standards~=2.0